requirements.txt
python-telegram-bot==13.4.1
matplotlib==3.3.3
pymongo==3.11.3
scipy==1.6.3
typing_extensions==3.7.4.3
requests==2.25.1